+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @under-control/core
 
+## 1.7.4
+
+### Patch Changes
+
+- Add nested error extraction mode
+
+## 1.7.3
+
+### Patch Changes
+
+- Add nested extractor
+
 ## 1.7.2
 
 ### Patch Changes